requirements.txt
PyYAML==6.0.2
web3==7.12.0
aio-pika==9.5.5
